packaging: remove cythonized files
See #13. Issue is that the cythonized .pyx file can stop matching
CPython APIs across Python versions, and that the .cpp file is
automatically used if it exists. This causes the Cython build to fail.

The gitignore change prevents .cpp files from being added in future
and adds a note about why not.